Fix:compile on Mac OS

+++ src/log.h
@@ -1,6 +1,7 @@
 #ifndef _LOG_H
 #define _LOG_H
 
+#include <stdint.h>
 #include "util.h"
 #include "platform.h"
 
@@ -10,12 +11,12 @@ struct log{
 	int fd;
 	int fd_db;
 	char name[LOG_NSIZE];
-	UINT db_alloc;
+	uint64_t db_alloc;
 	struct buffer *buf;
 };
 
 struct log *log_new(char *name);
-UINT log_append(struct log *log, struct slice *sk, struct slice *sv);
+uint64_t log_append(struct log *log, struct slice *sk, struct slice *sv);
 void log_trunc(struct log *log);
 void log_free(struct log *log);
