+++ src/skiplist.h
@@ -1,6 +1,6 @@
 #ifndef _SKIPLIST_H
 #define _SKIPLIST_H
-
+#include <stdint.h>
 #include "util.h"
 #include "platform.h"
 
@@ -11,7 +11,7 @@ typedef enum {ADD,DEL} OPT;
 
 struct skipnode{
     char key[SKIP_KSIZE];
-	UINT val;
+	uint64_t val;
 	OPT opt;                   
     struct skipnode *forward[1]; 
 	struct skipnode *next;
@@ -27,7 +27,7 @@ struct skiplist{
 };
 
 struct skiplist *skiplist_new(size_t size);
-int skiplist_insert(struct skiplist *list, struct slice *sk, UINT val, OPT opt);
+int skiplist_insert(struct skiplist *list, struct slice *sk, uint64_t val, OPT opt);
 struct skipnode *skiplist_lookup(struct skiplist *list, char *data);
 int skiplist_notfull(struct skiplist *list);
 void skiplist_dump(struct skiplist *list);
